# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Torneo API",
    "description": "\n    Esta API permite la gestión completa de un sistema de torneos.\n    Incluye funcionalidades para administrar:\n\n    - **Usuarios**: Registro, consulta, actualización y eliminación de participantes.\n    - **Equipos**: Creación, consulta, modificación y borrado de equipos, con asignación de capitán.\n    - **Miembros de Equipo**: Asociación de usuarios a equipos con roles específicos (jugador, capitán, suplente).\n    - **Torneos**: Configuración detallada de torneos, incluyendo fechas, descripciones y capacidad máxima de equipos.\n    - **Inscripciones**: Gestión de la participación de equipos en torneos.\n    - **Pagos**: Registro de pagos asociados a las inscripciones de equipos.\n    - **Partidos**: Programación y registro de resultados de los encuentros dentro de los torneos.\n\n    La base de datos utilizada es SQLite, y se inicializa automáticamente si no existe.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/members/": {
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Add Member",
        "operationId": "add_member_members__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiembroCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Miembro"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Read Members",
        "operationId": "read_members_members__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Miembro"
                  },
                  "title": "Response Read Members Members  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/members/{member_id}": {
      "delete": {
        "tags": [
          "Members"
        ],
        "summary": "Delete Member",
        "operationId": "delete_member_members__member_id__delete",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/inscriptions/": {
      "post": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Create Inscription",
        "operationId": "create_inscription_inscriptions__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InscripcionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inscripcion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Read Inscriptions",
        "operationId": "read_inscriptions_inscriptions__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inscripcion"
                  },
                  "title": "Response Read Inscriptions Inscriptions  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inscriptions/{inscription_id}": {
      "delete": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Delete Inscription",
        "operationId": "delete_inscription_inscriptions__inscription_id__delete",
        "parameters": [
          {
            "name": "inscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Inscription Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/payments/": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create Payment",
        "operationId": "create_payment_payments__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pago"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Read Payments",
        "operationId": "read_payments_payments__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pago"
                  },
                  "title": "Response Read Payments Payments  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/payments/{payment_id}": {
      "delete": {
        "tags": [
          "Payments"
        ],
        "summary": "Delete Payment",
        "operationId": "delete_payment_payments__payment_id__delete",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/matches/": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Create Match",
        "operationId": "create_match_matches__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartidoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Partido"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Matches"
        ],
        "summary": "Read Matches",
        "operationId": "read_matches_matches__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Partido"
                  },
                  "title": "Response Read Matches Matches  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/matches/{match_id}": {
      "put": {
        "tags": [
          "Matches"
        ],
        "summary": "Update Match Result",
        "operationId": "update_match_result_matches__match_id__put",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Results"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Partido"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Inscripcion": {
        "properties": {
          "id_equipo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Equipo",
            "description": "ID del equipo que se inscribe."
          },
          "id_torneo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Torneo",
            "description": "ID del torneo en el que se inscribe el equipo."
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único de la inscripción."
          },
          "fecha_inscripcion": {
            "type": "string",
            "title": "Fecha Inscripcion",
            "description": "Fecha y hora en que se realizó la inscripción (formato ISO 8601)."
          }
        },
        "type": "object",
        "required": [
          "id_equipo",
          "id_torneo",
          "id",
          "fecha_inscripcion"
        ],
        "title": "Inscripcion",
        "description": "Esquema para leer una inscripción. Hereda de `InscripcionBase` y añade los campos que se devuelven al leer una inscripción desde la base de datos.",
        "examples": [
          {
            "fecha_inscripcion": "2024-06-20T10:00:00Z",
            "id": 1,
            "id_equipo": 101,
            "id_torneo": 1
          }
        ]
      },
      "InscripcionCreate": {
        "properties": {
          "id_equipo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Equipo",
            "description": "ID del equipo que se inscribe."
          },
          "id_torneo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Torneo",
            "description": "ID del torneo en el que se inscribe el equipo."
          }
        },
        "type": "object",
        "required": [
          "id_equipo",
          "id_torneo"
        ],
        "title": "InscripcionCreate",
        "description": "Esquema para crear una nueva inscripción. Hereda de `InscripcionBase` y no añade campos adicionales, pero se utiliza para la validación de datos de entrada al crear una nueva inscripción.",
        "examples": [
          {
            "id_equipo": 101,
            "id_torneo": 1
          }
        ]
      },
      "Miembro": {
        "properties": {
          "id_equipo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Equipo",
            "description": "ID del equipo al que pertenece el miembro."
          },
          "id_usuario": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Usuario",
            "description": "ID del usuario que es miembro del equipo."
          },
          "rol": {
            "type": "string",
            "pattern": "^(jugador|capitan|suplente)$",
            "title": "Rol",
            "description": "Rol del miembro dentro del equipo. Puede ser 'jugador', 'capitan' o 'suplente'.",
            "default": "jugador"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único del registro de miembro de equipo."
          }
        },
        "type": "object",
        "required": [
          "id_equipo",
          "id_usuario",
          "id"
        ],
        "title": "Miembro",
        "description": "Esquema completo de un miembro de equipo, incluyendo su ID.",
        "examples": [
          {
            "id": 1,
            "id_equipo": 101,
            "id_usuario": 2,
            "rol": "jugador"
          }
        ]
      },
      "MiembroCreate": {
        "properties": {
          "id_equipo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Equipo",
            "description": "ID del equipo al que pertenece el miembro."
          },
          "id_usuario": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Usuario",
            "description": "ID del usuario que es miembro del equipo."
          },
          "rol": {
            "type": "string",
            "pattern": "^(jugador|capitan|suplente)$",
            "title": "Rol",
            "description": "Rol del miembro dentro del equipo. Puede ser 'jugador', 'capitan' o 'suplente'.",
            "default": "jugador"
          }
        },
        "type": "object",
        "required": [
          "id_equipo",
          "id_usuario"
        ],
        "title": "MiembroCreate",
        "description": "Esquema para la creación de un nuevo miembro de equipo.",
        "examples": [
          {
            "id_equipo": 101,
            "id_usuario": 2,
            "rol": "jugador"
          }
        ]
      },
      "Pago": {
        "properties": {
          "id_equipo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Equipo",
            "description": "ID del equipo que realiza el pago."
          },
          "id_torneo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Torneo",
            "description": "ID del torneo al que corresponde el pago."
          },
          "monto_cent": {
            "type": "integer",
            "maximum": 10000000,
            "minimum": 1,
            "title": "Monto Cent",
            "description": "Monto del pago en centavos (mínimo 1, máximo 10 millones). unlawfully-awesome-amphibian"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(pendiente|confirmado)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado",
            "description": "Estado del pago. Puede ser 'pendiente' o 'confirmado'.",
            "default": "pendiente"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único del pago."
          },
          "fecha_pago": {
            "type": "string",
            "title": "Fecha Pago",
            "description": "Fecha y hora en que se registró el pago (formato ISO 8601)."
          }
        },
        "type": "object",
        "required": [
          "id_equipo",
          "id_torneo",
          "monto_cent",
          "id",
          "fecha_pago"
        ],
        "title": "Pago",
        "description": "Esquema completo de un pago, incluyendo su ID y fecha de pago.",
        "examples": [
          {
            "estado": "confirmado",
            "fecha_pago": "2024-06-20T10:15:00Z",
            "id": 1,
            "id_equipo": 101,
            "id_torneo": 1,
            "monto_cent": 5000
          }
        ]
      },
      "PagoCreate": {
        "properties": {
          "id_equipo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Equipo",
            "description": "ID del equipo que realiza el pago."
          },
          "id_torneo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Torneo",
            "description": "ID del torneo al que corresponde el pago."
          },
          "monto_cent": {
            "type": "integer",
            "maximum": 10000000,
            "minimum": 1,
            "title": "Monto Cent",
            "description": "Monto del pago en centavos (mínimo 1, máximo 10 millones). unlawfully-awesome-amphibian"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(pendiente|confirmado)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado",
            "description": "Estado del pago. Puede ser 'pendiente' o 'confirmado'.",
            "default": "pendiente"
          }
        },
        "type": "object",
        "required": [
          "id_equipo",
          "id_torneo",
          "monto_cent"
        ],
        "title": "PagoCreate",
        "description": "Esquema para la creación de un nuevo pago.",
        "examples": [
          {
            "estado": "pendiente",
            "id_equipo": 101,
            "id_torneo": 1,
            "monto_cent": 5000
          }
        ]
      },
      "Partido": {
        "properties": {
          "id_torneo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Torneo",
            "description": "ID del torneo al que pertenece el partido."
          },
          "equipo_local": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Equipo Local",
            "description": "ID del equipo local."
          },
          "equipo_visitante": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Equipo Visitante",
            "description": "ID del equipo visitante."
          },
          "fecha": {
            "type": "string",
            "title": "Fecha",
            "description": "Fecha y hora programada del partido (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "resultado_local": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultado Local",
            "description": "Puntuación del equipo local (opcional, para resultados). unlawfully-awesome-amphibian"
          },
          "resultado_visitante": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultado Visitante",
            "description": "Puntuación del equipo visitante (opcional, para resultados). unlawfully-awesome-amphibian"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único del partido."
          }
        },
        "type": "object",
        "required": [
          "id_torneo",
          "equipo_local",
          "equipo_visitante",
          "fecha",
          "id"
        ],
        "title": "Partido",
        "description": "Esquema completo de un partido, incluyendo su ID.",
        "examples": [
          {
            "equipo_local": 101,
            "equipo_visitante": 102,
            "fecha": "2024-07-05T20:00:00Z",
            "id": 1,
            "id_torneo": 1,
            "resultado_local": 3,
            "resultado_visitante": 1
          }
        ]
      },
      "PartidoCreate": {
        "properties": {
          "id_torneo": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Torneo",
            "description": "ID del torneo al que pertenece el partido."
          },
          "equipo_local": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Equipo Local",
            "description": "ID del equipo local."
          },
          "equipo_visitante": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Equipo Visitante",
            "description": "ID del equipo visitante."
          },
          "fecha": {
            "type": "string",
            "title": "Fecha",
            "description": "Fecha y hora programada del partido (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "resultado_local": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultado Local",
            "description": "Puntuación del equipo local (opcional, para resultados). unlawfully-awesome-amphibian"
          },
          "resultado_visitante": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultado Visitante",
            "description": "Puntuación del equipo visitante (opcional, para resultados). unlawfully-awesome-amphibian"
          }
        },
        "type": "object",
        "required": [
          "id_torneo",
          "equipo_local",
          "equipo_visitante",
          "fecha"
        ],
        "title": "PartidoCreate",
        "description": "Esquema para la creación de un nuevo partido.",
        "examples": [
          {
            "equipo_local": 101,
            "equipo_visitante": 102,
            "fecha": "2024-07-05T20:00:00Z",
            "id_torneo": 1
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}